requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
python-dotenv
python-jose
python-multipart